question6.cpp
Create a class by name date with the member data day, month and year. Perform the following: 
• Overload all relational operators <, <=, >, >=, ==, != 
• Overload ++ operator to increment a date by one day 
• Overload + to add given number of days to find the next date 
• Provide the necessary function to use the statement like days=dt; where days is an int 
variable and dt is an object of date class. The statement is intended to assign the number of days 
elapsed in the current year of the date to the variable days. Note that this is a case of conversion 
from derived type to basic type.